New ELC and model updates.
Script updates in the ELC for additional item fields.
Simple diagnostic models for testing engine timing and input/output.

New firmware will be required for these tests. Look for firmware version 8534.
The EalEvent test will require new firmware for a bug in the injector event with the EAL Event item.